development_phase/backend/requirements.txt
aniso8601==9.0.1
bcrypt==4.0.0
click==8.1.3
Flask==2.2.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
ibm-db==3.1.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
PyJWT==2.5.0
python-dotenv==0.21.0
pytz==2022.4
six==1.16.0
Werkzeug==2.2.2
